runtime: String.prototype.repeat spec coercion + RangeError (no FPToSI-on-NaN)
Tier 3 of the code-smell audit (unguarded numeric coercion). repeat took
its count as int64 — the compiler FPToSI'd the JS double at the call site
(UB on NaN/Infinity → garbage count) and ts_string_repeat had no spec
guard, so `"x".repeat(-1)` silently returned "" (no RangeError),
`repeat(Infinity)` was UB, and `repeat(2**31)` attempted a multi-GB
allocation (OOM/hang).

Mirror the padStart/padEnd precedent: pass the raw double
(.f64Arg(ToF64) in the lowering registry + ts_string_repeat_native uses
ts_value_get_double) and apply ECMA-262 22.1.3.18 in the runtime —
n = ToIntegerOrInfinity(count); RangeError if n < 0 or n is +Infinity;
RangeError (not an allocation attempt) if the result would exceed the max
string length. NaN -> 0 -> "".

test262 +3 (27,774 -> 27,777), 0 regressions; golden-ir 267/279, node
295/297.

+++ src/runtime/include/TsString.h
@@ -120,7 +120,7 @@ extern "C" {
     void* ts_string_trimEnd(void* str);
     void* ts_string_substring(void* str, int64_t start, int64_t end);
     void* ts_string_slice(void* str, int64_t start, int64_t end);
-    void* ts_string_repeat(void* str, int64_t count);
+    void* ts_string_repeat(void* str, double count);
     void* ts_string_padStart(void* str, double targetLength, void* padString);
     void* ts_string_padEnd(void* str, double targetLength, void* padString);
     bool ts_string_startsWith(void* str, void* prefix);
